refactor: Improve smsdelivery ingredient interfaces
